+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/irs/IrsCallbackController.java
@@ -27,9 +27,12 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import jakarta.validation.constraints.Pattern;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.eclipse.tractusx.traceability.assets.domain.base.JobRepository;
+import org.eclipse.tractusx.traceability.assets.domain.base.OrderRepository;
+import org.eclipse.tractusx.traceability.assets.infrastructure.base.model.ProcessingState;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -38,16 +41,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @Hidden
 @RequiredArgsConstructor
+@Validated
 public class IrsCallbackController {
 
-    private final JobRepository jobRepository;
+    private final OrderRepository orderRepository;
 
     @Operation(operationId = "irsCallback",
-            summary = "Callback of irs get job details",
+            summary = "Callback of irs get order details",
             tags = {"IRSCallback"},
-            description = "The endpoint retrieves the information about a job which has been completed recently.",
+            description = "The endpoint retrieves the information about a order which has been completed recently.",
             security = @SecurityRequirement(name = "oAuth2", scopes = "profile email"))
-    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "Retrieves job id in completed state."),
+    @ApiResponses(value = {@ApiResponse(responseCode = "200", description = "Retrieves order id in completed state."),
             @ApiResponse(
                     responseCode = "400",
                     description = "Bad request.",
@@ -91,11 +95,15 @@ public class IrsCallbackController {
                     content = @Content(
                             mediaType = "application/json",
                             schema = @Schema(implementation = ErrorResponse.class)))})
-    @GetMapping("/irs/job/callback")
-    void handleIrsJobCallback(@RequestParam("id") String jobId, @RequestParam("state") String jobState) {
-        // Security measurment for injection
-        if (jobId.matches("^[a-zA-Z0-9_-]*$")) {
-            jobRepository.handleJobFinishedCallback(jobId, jobState);
-        }
+    @GetMapping("/irs/order/callback")
+    void handleIrsOrderCallback(
+            @RequestParam("orderId") @Pattern(regexp = "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
+                    message = "Invalid UUID format for orderId") String orderId,
+            @RequestParam(value = "batchId", required = false) @Pattern( regexp = "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-5][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
+                    message = "Invalid UUID format for batchId") String batchId,
+            @RequestParam("orderState") ProcessingState orderState,
+            @RequestParam(value = "batchState", required = false) ProcessingState batchState) {
+        orderRepository.handleOrderFinishedCallback(orderId, batchId, orderState, batchState);
     }
+
 }