+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/irs/OrderRepositoryImpl.java
@@ -1,7 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022, 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- * Copyright (c) 2022, 2023 ZF Friedrichshafen AG
- * Copyright (c) 2022, 2023, 2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2025 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -24,14 +22,16 @@ package org.eclipse.tractusx.traceability.assets.infrastructure.base.irs;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
-import org.eclipse.tractusx.traceability.assets.domain.base.JobRepository;
+import org.eclipse.tractusx.traceability.assets.domain.base.OrderRepository;
 import org.eclipse.tractusx.traceability.assets.domain.base.model.AssetBase;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.BomLifecycle;
-import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.RegisterJobRequest;
+import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.PartChainIdentificationKey;
+import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.RegisterOrderRequest;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.Direction;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.IRSResponse;
-import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.JobStatus;
+import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.IrsBatchResponse;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.factory.IrsResponseAssetMapper;
+import org.eclipse.tractusx.traceability.assets.infrastructure.base.model.ProcessingState;
 import org.eclipse.tractusx.traceability.common.properties.TraceabilityProperties;
 import org.eclipse.tractusx.traceability.contracts.domain.model.ContractAgreement;
 import org.eclipse.tractusx.traceability.contracts.domain.model.ContractType;
@@ -41,66 +41,79 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import static org.eclipse.tractusx.irs.component.enums.BomLifecycle.AS_BUILT;
 import static org.eclipse.tractusx.irs.component.enums.BomLifecycle.AS_PLANNED;
 
 @Slf4j
 @Service
-public class JobRepositoryImpl implements JobRepository {
+public class OrderRepositoryImpl implements OrderRepository {
 
     private final TraceabilityProperties traceabilityProperties;
     private final AssetCallbackRepository assetAsBuiltCallbackRepository;
     private final AssetCallbackRepository assetAsPlannedCallbackRepository;
-    private static final String JOB_STATUS_COMPLETED = "COMPLETED";
+    private static final ProcessingState STATUS_COMPLETED = ProcessingState.COMPLETED;
 
     private final IrsResponseAssetMapper assetMapperFactory;
 
-    private final JobClient jobClient;
+    private final OrderClient orderClient;
     private final ObjectMapper objectMapper;
+    private final JobClient jobClient;
 
-    public JobRepositoryImpl(
-            JobClient jobClient,
+
+    public OrderRepositoryImpl(
+            OrderClient orderClient,
             TraceabilityProperties traceabilityProperties,
             @Qualifier("assetAsBuiltRepositoryImpl")
             AssetCallbackRepository assetAsBuiltCallbackRepository,
             @Qualifier("assetAsPlannedRepositoryImpl")
             AssetCallbackRepository assetAsPlannedCallbackRepository,
-            IrsResponseAssetMapper assetMapperFactory, ObjectMapper objectMapper) {
+            IrsResponseAssetMapper assetMapperFactory,
+            ObjectMapper objectMapper, JobClient jobClient) {
         this.traceabilityProperties = traceabilityProperties;
         this.assetAsBuiltCallbackRepository = assetAsBuiltCallbackRepository;
         this.assetAsPlannedCallbackRepository = assetAsPlannedCallbackRepository;
-        this.jobClient = jobClient;
+        this.orderClient = orderClient;
         this.assetMapperFactory = assetMapperFactory;
         this.objectMapper = objectMapper;
+        this.jobClient = jobClient;
     }
 
     @Override
-    public void createJobToResolveAssets(String globalAssetId, Direction direction, List<String> aspects, BomLifecycle bomLifecycle) {
-        RegisterJobRequest registerJobRequest = RegisterJobRequest.buildJobRequest(globalAssetId, traceabilityProperties.getBpn().toString(), direction, aspects, bomLifecycle, traceabilityProperties.getUrl());
-        this.jobClient.registerJob(registerJobRequest);
+    public void createOrderToResolveAssets(List<String> globalAssetIds, Direction direction, List<String> aspects, BomLifecycle bomLifecycle) {
+        final String callbackUrl = traceabilityProperties.getUrl();
+        String applicationBpn = traceabilityProperties.getBpn().toString();
+        List<PartChainIdentificationKey> keys = globalAssetIds.stream().map(s -> new PartChainIdentificationKey(s, applicationBpn)).collect(Collectors.toList());
+        RegisterOrderRequest registerOrderRequest = RegisterOrderRequest.buildOrderRequest(aspects, bomLifecycle, callbackUrl, direction, keys);
+        this.orderClient.registerOrder(registerOrderRequest);
     }
 
 
     @Override
-    public void handleJobFinishedCallback(String jobId, String state) {
-        if (!Objects.equals(state, JOB_STATUS_COMPLETED)) {
+    public void handleOrderFinishedCallback(String orderId, String batchId, ProcessingState orderState, ProcessingState batchState) {
+
+        if (!STATUS_COMPLETED.equals(batchState)) {
             return;
         }
 
-        final IRSResponse jobResponseIRS = this.jobClient.getIrsJobDetailResponse(jobId);
+        final IrsBatchResponse irsBatchResponse = this.orderClient.getBatchByOrder(orderId, batchId);
 
-        if (jobResponseIRS == null) {
+        if (irsBatchResponse == null) {
+            log.warn("Received order callback by the item relationship service is incorrect");
             return;
         }
 
-        long runtime = (jobResponseIRS.jobStatus().lastModifiedOn().getTime() - jobResponseIRS.jobStatus().startedOn().getTime()) / 1000;
-        log.info("IRS call for globalAssetId: {} finished with status: {}, runtime {} s.", jobResponseIRS.jobStatus().globalAssetId(), jobResponseIRS.jobStatus().state(), runtime);
 
-        if (jobCompleted(jobResponseIRS.jobStatus())) {
-            List<AssetBase> assets = assetMapperFactory.toAssetBaseList(jobResponseIRS);
+        if (irsBatchResponse.jobs().isEmpty()) {
+            log.warn("Received order callback by the item relationship service has no jobs.");
+            return;
+        }
+
+        irsBatchResponse.jobs().forEach(jobRecord -> {
+            IRSResponse irsJobDetailResponse = jobClient.getIrsJobDetailResponse(jobRecord.id());
+            List<AssetBase> assets = assetMapperFactory.toAssetBaseList(irsJobDetailResponse);
             assets.forEach(assetBase -> {
                 if (assetBase.getBomLifecycle() == AS_BUILT) {
                     saveOrUpdateAssets(assetAsBuiltCallbackRepository, assetBase);
@@ -108,7 +121,8 @@ public class JobRepositoryImpl implements JobRepository {
                     saveOrUpdateAssets(assetAsPlannedCallbackRepository, assetBase);
                 }
             });
-        }
+        });
+
     }
 
     void saveOrUpdateAssets(AssetCallbackRepository repository, AssetBase asset) {
@@ -142,10 +156,6 @@ public class JobRepositoryImpl implements JobRepository {
         } catch (JsonProcessingException e) {
             throw new RuntimeException(e);
         }
-    }
-
-    public static boolean jobCompleted(JobStatus jobStatus) {
-        return JOB_STATUS_COMPLETED.equals(jobStatus.state());
     }
 
 }