+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/infrastructure/base/irs/model/response/IrsBatchResponse.java
@@ -1,5 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2025 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -16,18 +16,28 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
+package org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response;
 
-package org.eclipse.tractusx.traceability.submodel.domain.model;
+import java.util.Date;
+import java.util.List;
 
-
-import lombok.Builder;
-import lombok.Data;
-
-import java.util.Map;
-
-@Data
-@Builder
-public class SubmodelRequest {
-    private String submodelId;
-    private Map<String, Object> data;
+public record IrsBatchResponse(
+        String batchId,
+        int batchNumber,
+        String batchProcessingState,
+        int batchTotal,
+        Date completedOn,
+        List<JobRecord> jobs,
+        int jobsInBatchChecksum,
+        String orderId,
+        Date startedOn,
+        int totalJobs
+) {
+    public record JobRecord(
+            Date completedOn,
+            String id,
+            Date startedOn,
+            String state
+    ) {
+    }
 }