+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/assets/domain/service/AssetAsPlannedServiceImplTest.java
@@ -21,7 +21,7 @@ package org.eclipse.tractusx.traceability.assets.domain.service;
 
 import org.eclipse.tractusx.traceability.assets.domain.asplanned.repository.AssetAsPlannedRepository;
 import org.eclipse.tractusx.traceability.assets.domain.asplanned.service.AssetAsPlannedServiceImpl;
-import org.eclipse.tractusx.traceability.assets.domain.base.JobRepository;
+import org.eclipse.tractusx.traceability.assets.domain.base.OrderRepository;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.request.BomLifecycle;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.Direction;
 import org.eclipse.tractusx.traceability.assets.infrastructure.base.irs.model.response.relationship.Aspect;
@@ -30,6 +30,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.util.List;
 
 import static org.mockito.Mockito.verify;
 
@@ -40,7 +42,7 @@ class AssetAsPlannedServiceImplTest {
     private AssetAsPlannedServiceImpl assetService;
 
     @Mock
-    private JobRepository jobRepository;
+    private OrderRepository orderRepository;
 
     @Mock
     private AssetAsPlannedRepository assetRepository;
@@ -49,12 +51,12 @@ class AssetAsPlannedServiceImplTest {
     void synchronizeAssets_shouldSaveCombinedAssets_whenNoException() {
         // given
         String globalAssetId = "123";
-
+        List<String> globalAssetIds = List.of(globalAssetId);
         // when
-        assetService.synchronizeAssetsAsync(globalAssetId);
+        assetService.syncAssetsAsyncUsingIRSOrderAPI(globalAssetIds);
 
         // then
-        verify(jobRepository).createJobToResolveAssets(globalAssetId, Direction.DOWNWARD, Aspect.downwardAspectsForAssetsAsPlanned(), BomLifecycle.AS_PLANNED);
+        verify(orderRepository).createOrderToResolveAssets(globalAssetIds, Direction.DOWNWARD, Aspect.downwardAspectsForAssetsAsPlanned(), BomLifecycle.AS_PLANNED);
     }
 }