+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/assets/infrastructure/base/IrsCallbackControllerIT.java
@@ -22,6 +22,7 @@ package org.eclipse.tractusx.traceability.integration.assets.infrastructure.base
 import assets.importpoc.ImportResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.restassured.http.ContentType;
+import org.apache.http.protocol.HTTP;
 import org.eclipse.tractusx.traceability.common.security.JwtRole;
 import org.eclipse.tractusx.traceability.integration.IntegrationTestSpecification;
 import org.eclipse.tractusx.traceability.integration.common.support.AssetsSupport;
@@ -38,6 +39,7 @@ import java.io.File;
 
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.eclipse.tractusx.traceability.integration.common.support.IrsApiSupport.sendCallback;
 
 class IrsCallbackControllerIT extends IntegrationTestSpecification {
 
@@ -56,6 +58,8 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
     @Autowired
     BpnSupport bpnSupport;
 
+    private final static int HTTP_STATUS_OKAY = 200;
+
     @BeforeEach
     void setUp() throws JsonProcessingException {
         edcSupport.performSupportActionsForBpdmAccess();
@@ -68,39 +72,21 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
             throws JoseException {
         // given
         oAuth2ApiSupport.oauth2ApiReturnsTechnicalUserToken();
+        irsApiSupport.irsApiReturnsOrderAndBatchDetails();
         irsApiSupport.irsApiReturnsJobDetails();
 
-        String jobId = "ebb79c45-7bba-4169-bf17-3e719989ab54";
-        String jobState = "COMPLETED";
+        String orderId = "ebb79c45-7bba-4169-bf17-3e719989ab54";
+        String batchId = "ebb79c45-7bba-4169-bf17-3e719989ab55";
+        String orderState = "PROCESSING";
+        String batchState = "COMPLETED";
 
         // when
-        given()
-                .contentType(ContentType.JSON)
-                .log().all()
-                .when()
-                .param("id", jobId)
-                .param("state", jobState)
-                .get("/api/irs/job/callback")
-                .then()
-                .log().all()
-                .statusCode(200);
+        sendCallback(orderId, batchId, orderState, batchState, HTTP_STATUS_OKAY);
 
         // then
         assertThat(bpnSupportRepository.findAll()).hasSize(1);
         assetsSupport.assertAssetAsBuiltSize(2);
         assetsSupport.assertAssetAsPlannedSize(0);
-
-        // Make the API call and store the response
-        given()
-                .header(oAuth2Support.jwtAuthorization(JwtRole.ADMIN))
-                .contentType(ContentType.JSON)
-                .log().all()
-                .when()
-                .pathParam("assetId", "urn:uuid:d387fa8e-603c-42bd-98c3-4d87fef8d2bb")
-                .get("/api/assets/as-built/{assetId}")
-                .then()
-                .log().all()
-                .statusCode(200);
     }
 
     @Test
@@ -108,84 +94,59 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
         // given
         oAuth2ApiSupport.oauth2ApiReturnsTechnicalUserToken();
         irsApiSupport.irsJobDetailsAsPlanned();
-        String jobId = "ebb79c45-7bba-4169-bf17-SUCCESSFUL_AS_PLANNED";
-        String jobState = "COMPLETED";
-
+        irsApiSupport.irsApiReturnsOrderAndBatchDetailsAsPlanned();
+        String orderId = "ebb79c45-7bba-4169-bf17-3e719989ab57";
+        String batchId = "ebb79c45-7bba-4169-bf17-3e719989ab58";
+        String orderState = "COMPLETED";
+        String batchState = "COMPLETED";
         // when
-        given()
-                .contentType(ContentType.JSON)
-                .log().all()
-                .when()
-                .param("id", jobId)
-                .param("state", jobState)
-                .get("/api/irs/job/callback")
-                .then()
-                .log().all()
-                .statusCode(200);
+        sendCallback(orderId, batchId, orderState, batchState, HTTP_STATUS_OKAY);
 
         // then
-        assertThat(bpnSupportRepository.findAll()).hasSize(1);
-        assetsSupport.assertAssetAsBuiltSize(0);
         assetsSupport.assertAssetAsPlannedSize(3);
     }
 
     @Test
-    void givenInvalidJobId_whenCallbackReceivedForAsPlanned_thenNothingSynchronized() {
+    void givenNoBatchIdNoBatchState_whenCallbackReceivedOrderCompleted_thenSimplyDoNothing() {
         // given
         oAuth2ApiSupport.oauth2ApiReturnsTechnicalUserToken();
         irsApiSupport.irsJobDetailsAsPlanned();
-        String invalidJobId = "irs/admin/test/ID";
-        String jobState = "COMPLETED";
+        irsApiSupport.irsApiReturnsOrderAndBatchDetailsAsPlanned();
+        String orderId = "ebb79c45-7bba-4169-bf17-3e719989ab57";
+        String batchId = "ebb79c45-7bba-4169-bf17-3e719989ab58";
+        String orderState = "COMPLETED";
 
-        // when
+        // when then
         given()
                 .contentType(ContentType.JSON)
                 .log().all()
                 .when()
-                .param("id", invalidJobId)
-                .param("state", jobState)
-                .get("/api/irs/job/callback")
+                .param("orderId", orderId)
+                .param("orderState", orderState)
+                .get("/api/irs/order/callback")
                 .then()
                 .log().all()
-                .statusCode(200);
+                .statusCode(HTTP_STATUS_OKAY);
+
+    }
+
+
+    @Test
+    void givenNoJobsForOrder_whenCallbackReceivedForAsPlanned_thenNothingSynchronized() {
+        // given
+        oAuth2ApiSupport.oauth2ApiReturnsTechnicalUserToken();
+        irsApiSupport.irsApiReturnsOrderAndBatchDetailsAsPlannedNoJobs();
+        String orderId = "ebb79c45-7bba-4169-bf17-3e719989ab51";
+        String batchId = "ebb79c45-7bba-4169-bf17-3e719989ab52";
+        String orderState = "COMPLETED";
+        String batchState = "COMPLETED";
+
+        // when
+        sendCallback(orderId, batchId, orderState, batchState, HTTP_STATUS_OKAY);
 
         // then
         assertThat(bpnSupportRepository.findAll()).isEmpty();
-        assetsSupport.assertAssetAsBuiltSize(0);
         assetsSupport.assertAssetAsPlannedSize(0);
-    }
-
-    @Test
-    void givenAssetExist_whenCallbackReceived_thenUpdateIt() {
-        // given
-        assetsSupport.defaultAssetsStored();
-        oAuth2ApiSupport.oauth2ApiReturnsTechnicalUserToken();
-        irsApiSupport.irsApiReturnsJobDetails();
-        String jobId = "ebb79c45-7bba-4169-bf17-3e719989ab54";
-        String jobState = "COMPLETED";
-
-        // when
-        given()
-                .contentType(ContentType.JSON)
-                .log().all()
-                .when()
-                .param("id", jobId)
-                .param("state", jobState)
-                .get("/api/irs/job/callback")
-                .then()
-                .log().all()
-                .statusCode(200);
-
-
-        // then
-        assertThat(bpnSupportRepository.findById("BPNL000000012345"))
-                .hasValueSatisfying(entity -> assertThat(entity.getManufacturerName()).isEqualTo("OEM A Short"));
-        assetsSupport.assertAssetAsBuiltSize(14);
-        assetsSupport.assertAssetAsPlannedSize(0);
-        String updatedIdShort = assetsSupport.findById("urn:uuid:6dafbcec-2fce-4cbb-a5a9-b3b32aa5cffc").getIdShort();
-        assertThat(updatedIdShort).isEqualTo("ecu.asm");
-        //urn:uuid:d387fa8e-603c-42bd-98c3-4d87fef8d2bb
-        //vehicle_hybrid_v2.asm
     }
 
     @Test
@@ -194,8 +155,11 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
 
         oAuth2ApiSupport.oauth2ApiReturnsTechnicalUserToken();
         irsApiSupport.irsApiReturnsJobDetails();
-        String jobId = "ebb79c45-7bba-4169-bf17-3e719989ab54";
-        String jobState = "COMPLETED";
+        irsApiSupport.irsApiReturnsOrderAndBatchDetails();
+        String orderId = "ebb79c45-7bba-4169-bf17-3e719989ab54";
+        String batchId = "ebb79c45-7bba-4169-bf17-3e719989ab55";
+        String orderState = "COMPLETED";
+        String batchState = "COMPLETED";
 
         String path = getClass().getResource("/testdata/importfiles/validImportFile.json").getFile();
         File file = new File(path);
@@ -206,20 +170,11 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
                 .multiPart(file)
                 .post("/api/assets/import")
                 .then()
-                .statusCode(200)
+                .statusCode(HTTP_STATUS_OKAY)
                 .extract().as(ImportResponse.class);
 
         // when
-        given()
-                .contentType(ContentType.JSON)
-                .log().all()
-                .when()
-                .param("id", jobId)
-                .param("state", jobState)
-                .get("/api/irs/job/callback")
-                .then()
-                .log().all()
-                .statusCode(200);
+        sendCallback(orderId, batchId, orderState, batchState, HTTP_STATUS_OKAY);
 
         // then
         String tombstoneAsBuilt = given()
@@ -231,7 +186,7 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
                 .get("/api/assets/as-built/{assetId}")
                 .then()
                 .log().all()
-                .statusCode(200)
+                .statusCode(HTTP_STATUS_OKAY)
                 .extract().path("tombstone");
 
         assertThat(tombstoneAsBuilt).isNotEmpty();
@@ -243,8 +198,11 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
         // given
         oAuth2ApiSupport.oauth2ApiReturnsTechnicalUserToken();
         irsApiSupport.irsJobDetailsAsPlanned();
-        String jobId = "ebb79c45-7bba-4169-bf17-SUCCESSFUL_AS_PLANNED";
-        String jobState = "COMPLETED";
+        irsApiSupport.irsApiReturnsOrderAndBatchDetailsAsPlanned();
+        String orderId = "ebb79c45-7bba-4169-bf17-3e719989ab57";
+        String batchId = "ebb79c45-7bba-4169-bf17-3e719989ab58";
+        String orderState = "COMPLETED";
+        String batchState = "COMPLETED";
 
         String path = getClass().getResource("/testdata/importfiles/validImportFile-onlyAsPlannedAsset.json").getFile();
         File file = new File(path);
@@ -255,20 +213,11 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
                 .multiPart(file)
                 .post("/api/assets/import")
                 .then()
-                .statusCode(200)
+                .statusCode(HTTP_STATUS_OKAY)
                 .extract().as(ImportResponse.class);
 
         // when
-        given()
-                .contentType(ContentType.JSON)
-                .log().all()
-                .when()
-                .param("id", jobId)
-                .param("state", jobState)
-                .get("/api/irs/job/callback")
-                .then()
-                .log().all()
-                .statusCode(200);
+        sendCallback(orderId, batchId, orderState, batchState, HTTP_STATUS_OKAY);
 
         // then
         String tombstoneAsPlanned = given()
@@ -280,7 +229,7 @@ class IrsCallbackControllerIT extends IntegrationTestSpecification {
                 .get("/api/assets/as-planned/{assetId}")
                 .then()
                 .log().all()
-                .statusCode(200)
+                .statusCode(HTTP_STATUS_OKAY)
                 .extract().path("tombstone");
 
         assertThat(tombstoneAsPlanned).isNotEmpty();